Add a direct dependency on cluster_id before deciding it's ready
With `count=0` on some Fargate resources, there's no dependency that prevents, say, helm_releases from attempting to apply/destroy immediately, even when the cluster's not healthy and nodes aren't available. So we add both a direct dependency on the availability of both the eks.cluster_id, and the availability of the managed node groups, which take longer to come up.