fix: break circular/redundant dependencies for kubernetes resources
Now that we have managed node groups, we don't need to patch coredns to enable it to run on Fargate. So we remove that provisioner.

Since we are now blocking `kubernetes provider` -> `null_resource.cluster-functional` -> `module.eks.cluster_id (and node groups)`, we've eliminated the possibility that kubernetes interactions happen before the cluster is ready. So we can also remove explicit dependencies on null_resource.cluster-functional where they're no longer needed, and let Terraform infer the dependencies.