Add support for Bind Module Reference (#79)
* update: terraform/bind/versions.tf

Getting error 'Module module.brokerpak-eks-terraform-bind contains provider configuration' in ssb broker

* update: match aws version with provision and bind

* fix\!: modules can't depend on modules with providers

Module on module dependency does not allow
 the child module to specify its own provider blocks; all provider blocks must be in the root module

* new: add explicit ready output for binding to wait on

* new: use AWS_PROFILE instead of env vars to aws credentials

There's overlap between different credentials in our deployment, so use AWS_PROFILE to have finer control over which credentials are used when.. Reference: https://docs.aws.amazon.com/sdk-for-go/v1/developer-guide/configuring-sdk.html#specifying-credentials

* fix: env is an argument, not a block

* new: add variable to accept aws credentials

Since the module is being run directly from the same context as the ssb, we have to set different credentials for the space in which eks will live (which is different from the ssb backend)

* new: child modules are not supposed to manage providers

This is a shift in how we are using EKS terraform from brokerpak-only to brokerpak AND terraform module

* cleanup: remove aws credential variable requirement

* test: add aws provider back for refresh issue

To fix 'To work with module.brokerpak-eks-terraform-provision.aws_kms_key.cluster its original provider configuration at
│ module.brokerpak-eks-terraform-provision.provider[registry.terraform.io/hashicorp/aws].dnssec-key-provider is required, but it has been removed. This occurs when a
│ provider configuration is removed while objects created by that provider still exist in the state. Re-add the provider configuration to destroy
│ module.brokerpak-eks-terraform-provision.aws_kms_key.cluster, after which you can remove the provider configuration again.'

* revert: remove aws provider from sub-module

Fix for.. Error: error configuring Terraform AWS Provider: error validating provider credentials: error calling sts:GetCallerIdentity: InvalidClientTokenId: The security token included in the request is invalid.
│       status code: 403, request id: 38dd9ba8-75e8-4b9d-93c5-dac8c5a2d534
│
│   with module.brokerpak-eks-terraform-provision.provider[registry.terraform.io/hashicorp/aws].dnssec-key-provider,
│   on .terraform/modules/brokerpak-eks-terraform-provision/terraform/provision/providers.tf line 2, in provider aws:
│    2: provider aws {

* new: declare require_providers for module

Child modules must specify required_providers and use their defined local name (if different from the providers preferred name); Reference: https://www.terraform.io/language/providers/requirements#handling-local-name-conflicts

* fix: remove vestigial code from the previous approach

* fix: properly require an aliased aws provider to the module

See https://www.terraform.io/language/modules/develop/providers#provider-aliases-within-modules

* refactor: manage fargate profiles via the EKS module support

Now that we're using the up-to-date version of the EKS module, the module is better at managing Fargate profiles than we are. So now we let it handle that.

* refactor: setup logging using explicit kubernetes resources

We were previously restricted to using a version of the kubernetes provider that didn't support annotations, so we were forced to use a null_resource to set up logging. Now that we're no longer restricted, we can do this with explicit namespace and configmap resources.

* fix: ensure all pods are logged, whether in nodegroups or in Fargate

Ensure this policy is applied to the roles created by the EKS module for all node_groups and fargate_profiles. This way all pod logs go to Cloudwatch.

(The policy name were also changed, since it's not only for use with Fargate pods.)

* refactor: ensure all EKS module-created IAM roles are able to use EBS

* fix: require AWS creds to use with the kubernetes provider

The kubernetes provider needs AWS creds set in the environment so it can use aws-iam-authenticator to connect.

* fix: remove extraneous brace

* refactor: provision and bind are relocated and treated as submodules

By making provision and bind into submodules of a single root module that encompasses them both (in ../terraform), we can hoist their provider configurations upward into the root module. Then the root module can handle accepting the provider configuration from callers (in this case datagov-ssb's static provision+bind deployment) and pass them to the submodules. We can have the bind module depend on the provision module, which skirts Terraform restrictions that would normally prevent the bind module from declaring its own kubernetes provider.

The submodules remain usable as root-modules in other contexts. For example, the brokerpak can use provision once, then use bind multiple times for each binding.

* fix: bind requires AWS creds to set up its kubernetes provider

We still can't yet uncomment/declare the bind module because of the depends_on relationship with provision.

The next step is to try to hoist the kubernetes provider out of the bind module. Ideally we'll be able to specify a kubeconfig directly using the output of the provision module.

* refactor: hoist provision's kubernetes provider config to root module

* refactor: hoist bind's kubernetes provider config up to the root module

* feature: explicitly widen compatibility to include anything before Terraform 2.x

* clarity: make scope of namespace-admin role explicit in the name

* refactor: provision module generates token for a cluster-admin service account

The token is used in the kubernetes provider config for bind, removing the need for any AWS credentials or binaries.

* Comment out EKS module's fargate_profiles

With FargateProfiles configured, the ebs-csi addon and the ingress-nginx helm release never start up.

* Explicitly inject the configured helm provider

* Turn on debugging for the helm provider

* refactor/fix: update location of terraform code for brokerpak

Due to previous refactor, the location of the provision/bind/provider terraform code was split up;  this updates it for the brokerpak to still know where things are

* refactor/fix: add new outputs for brokerpak provision

Due to the additional dependencies between provision and binding, provisioning has more outputs for the brokerpak to track now

* refactor/fix: add AWS Creds to brokerpak provision input

Brokerpak was complaining that fields were being used without being declared.

* fix: helm provider should use the exec clause to get tokens

* fix: bind doesn't need an aws provider

* cleanup: remove outdated comments

* refactor: use the same output/input names as one would find in a kubeconfig

* new: add provider configuration only used while brokering

When brokering, these are top-level modules, so they need to do their own provider setup

* cleanup: remove unused output

* cleanup: remove unused input, output, and comments

* fix: AWS creds only need to be provided in the non-brokered (root-level module) case

* fix: The brokerpak needs to include the new provision/admin-account.tf

* fix: bubble outputs up out of the submodules

* fix: confirm we do want to output the sensitive token value

* fix: add implicit dependency between logging configmap and namespace

* fix: be explicit about wanting to output a sensitive token

* feat: make node groups distinguishable in the AWS Console

* feat: name policies for same k8s cluster to sort in console

* fix: leave extra time for MNGs to destroy

* cleanup: ran `terraform fmt`

* fix: test fixture path updated

Co-authored-by: Bret Mogilefsky <bret.mogilefsky@gsa.gov>