fix: use a heredoc to specify a subdomain for the text fixture
We want to test an ingress with a subdomain, because that way we know that external-dns is populating DNS records.

In order for testing through SSL to work, we also need the full subdomain in the Ingress definition. By using a heredoc we can use a variable right in the script to specify that full name.