fix: avoid removing any AWS-side resources before k8s-side resources
This helps account for undocumented dependencies between helm_releases and AWS resources.

NOTE: This may slow down provision/deprovision of the entire stack; so be it!

If we need to optimize in future we can remove this and add explicit dependencies where we can identify that they're needed (eg for the alb controller, which is unable to remove LBs it provisioned without this change.