fix: ensure helm_release.calico uninstalls cleanly
Without this code, on `terraform destroy` we see "Error:
uninstallation completed with 1 error(s): uninstall: Failed to purge the
release: release: not found".

This change introduces a direct dependency between code in the provision-aws and
provision-k8s modules. If these modules are to remain usable separately in future,
then we need to identify the specific resource that helm_release.calico depends on
and find another way to make the dependency explicit, eg via something passed in
via a variable.