feat: this is pure genius (I think)
The issue was not about anything cluster specific.  The real issue was that the machine where terraform was being run couldn't access the control plane because the control plane was denying everything except what it was configured with.  So, terraform was able to lookup the control plane server ip, but it was not able to communicate with it.  This is a nice solution inspired by: https://stackoverflow.com/questions/46763287/i-want-to-identify-the-public-ip-of-the-terraform-execution-environment-and-add