Restrict acess to EKS Control Plane (#101)
* refactor: organize terraform variables

Document which variables are required, optional or operationally impactful

* new: implement cluster_endpoint_public_access_cidrs

- Configure cluster_endpoint_public_access_cidrs to be a terraform variable
- Specify terraform variable in broker specification
- Grab the Host IP of the machine starting the broker to be able to run the tests on github actions
- Default cluster_endpoint_public_access_cidrs to be the cloud.gov egress ranges

* lint: terraform fmt

* fix: also allow internal vpc traffic

Didn't test locally, just hoping it works on github :s

* reimplement: same idea, just implemented better

less error prone.. append internal vpc cide in the code as opposed to as configuration

* fix: try NAT!

okay internal ips can't connect externally, so let's try this instead

* new/fix: IPs for CRDs need to be allowed

This is.. well.. something I don't like

* new: the errors are from EC2 node IPs

There are a lot of IPs in the us-west-2 region and the reason the other tests failed was because the EC2 nodes (where the worker nodes/control plane nodes lived) weren't able to talk itself... this is a blanket to see if it works, but there should be a way to narrow it down... I fear that the EC2 nodes are actually the load balancer nodes which aren't known until the EKS cluster is already up and therefore can't be modified.. this is all speculation...

* new: create additional allowance after lb creation

This breaks a lot of molds around how this is all configured and dependent.  I'm testing whether terraform can figure out dependencies properly before I add any in.  Solution was inspired from: https://stackoverflow.com/questions/56713493/terraform-how-to-get-ip-address-of-aws-lb and validated against these references,
- https://registry.terraform.io/modules/terraform-aws-modules/eks/aws/latest?tab=outputs
- https://registry.terraform.io/modules/terraform-aws-modules/vpc/aws/latest?tab=outputs
- https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/security_group
- https://registry.terraform.io/providers/hashicorp/aws/latest/docs/data-sources/network_interface
- https://github.com/terraform-aws-modules/terraform-aws-eks/blob/dc8a6eecddc2c6957ba309a939ee46f39b946461/main.tf
- https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/eks_cluster
- https://docs.aws.amazon.com/eks/latest/userguide/sec-group-reqs.html

* feat: this is pure genius (I think)

The issue was not about anything cluster specific.  The real issue was that the machine where terraform was being run couldn't access the control plane because the control plane was denying everything except what it was configured with.  So, terraform was able to lookup the control plane server ip, but it was not able to communicate with it.  This is a nice solution inspired by: https://stackoverflow.com/questions/46763287/i-want-to-identify-the-public-ip-of-the-terraform-execution-environment-and-add

* new/fix: add provider configuration for brokerpak

http is a new provider

* fix: dependencies..

not sure why these things didn't depend on cluster-functional before, but it seems like it fixes the problem :/

* docs: update info about control_plane_ingress_cidrs

Fix the config in makefile

* docs: update missing info about single_az