new: convert autocaler-provisioner to cluster-autoscaler
This looks to be working.  It is an upgrade because the autoscaler-provisioner is deprecated.  It might not be working because of fargate configuration that still needs some fine-tuning (in terms of spinning up new nodes).  However, the helm chart deploys.  I haven't hit a blocker that is definitive that we can't use fargate.  This requires more investigation.
References:
- https://github.com/kubernetes/autoscaler/blob/master/cluster-autoscaler/cloudprovider/aws/README.md#auto-discovery-setup
- https://github.com/kubernetes/autoscaler/blob/master/charts/cluster-autoscaler/Chart.yaml
- https://artifacthub.io/packages/helm/cluster-autoscaler/cluster-autoscaler